requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
decorator==4.4.2
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
kombu==5.2.4
moviepy==1.0.3
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
